Extracting IRqliteClient to a separate file
